Added functional tests for student/teacher login and logout and changed header icon to use material design
