+++ peerinst/templates/peerinst/student/header.html
@@ -9,7 +9,7 @@
 
 {% else %}
 
-<i class="icon ion-md-menu header-icon"></i>
+<i id="menu-icon" class="material-icons md-32 header-icon">menu</i>
 <aside class="mdc-drawer mdc-drawer--temporary mdc-typography">
   <nav class="mdc-drawer__drawer">
