+++ peerinst/templates/peerinst/teacher/header.html
@@ -1,6 +1,6 @@
 {% load staticfiles i18n %}
 
-<i class="icon ion-md-menu header-icon"></i>
+<i id="menu-icon" class="material-icons md-32 header-icon">menu</i>
 <aside class="mdc-drawer mdc-drawer--temporary mdc-typography">
   <nav class="mdc-drawer__drawer">
     <nav id="icon-with-text-demo" class="mdc-drawer__content mdc-list">
